backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.38.0
python-multipart==0.0.21
email-validator==2.2.0   
redis==5.2.0
celery==5.6.2
boto3==1.42.25
python-dotenv==1.2.1
pydantic-settings==2.12.0
sqlalchemy==2.0.45
pyjwt==2.10.1
stripe==14.1.0
alembic==1.18.0
psycopg2-binary==2.9.11
pgvector==0.4.2
sentence-transformers==5.2.0